pyproject.toml
[tool.poetry]
name = "gs-videoreport"
version = "0.2.0"
description = "AI-powered video-to-lesson-plan converter for teachers"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "gs_video_report", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
google-genai = "^0.3.0"  # 最新 Google Gen AI SDK
typer = "^0.12.0"        # CLI 框架
pyyaml = "^6.0"          # 配置文件解析
yt-dlp = "^2025.1.26"    # YouTube 视频下载
httpx = "^0.27.0"        # HTTP 客户端 (google-genai 依赖)
rich = "^13.0.0"         # Rich CLI output
requests = "^2.31.0"     # HTTP requests for URL validation
aiohttp = {version = "^3.9.0", optional = true}  # 异步HTTP客户端（可选）

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"        # 测试框架
pytest-asyncio = "^0.23.0"  # 异步测试支持
pytest-mock = "^3.14.0"     # Mock 支持
black = "^24.0.0"            # 代码格式化
isort = "^5.13.0"            # Import 排序
mypy = "^1.8.0"              # 类型检查

[tool.poetry.scripts]
gs_videoreport = "gs_video_report.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 可选：如果需要更快的异步性能
[tool.poetry.extras]
fast = ["aiohttp"]
